--- google-cloud-video-stitcher/samples/acceptance/cloud_cdn_key_definition.rb
@@ -1,12 +1,0 @@
-require "google/cloud/video/stitcher"
-
-def cloud_cdn_def cdn_key_path, hostname, key_name, private_key
-  {
-    name: cdn_key_path,
-    hostname: hostname,
-    google_cdn_key: {
-      key_name: key_name,
-      private_key: private_key
-    }
-  }
-end
