--- google-cloud-video-stitcher/samples/acceptance/slate_definition.rb
@@ -1,7 +1,0 @@
-require "google/cloud/video/stitcher"
-
-def slate_def slate_uri
-  {
-    uri: slate_uri
-  }
-end
